helpdesk_proj/requirements.txt
Django==1.6.2
backports.ssl-match-hostname==3.4.0.2
django-browserid==0.9
djangorestframework==2.3.12
fancy-tag==0.2.0
gunicorn==18.0
requests==2.2.1
simplejson==3.3.3
six==1.5.2
tornado==3.2
wsgiref==0.1.2
